+++ CHANGES.md
@@ -1,4 +1,6 @@
 [unreleased]
+
+#### 1.1.0 / 2022-06-02
 * change 'Activate' plugin action link to 'Cannot Activate' text when plugin has unmet dependencies
 * remove checkbox from plugin row when plugin has unmet dependencies
 * use _View details_ link for plugins listed in **Requires:** in plugin row